begin escape char for quotes - update tests

+++ includes/lexer.h
@@ -6,7 +6,7 @@
 /*   By: jjaniec <jjaniec@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/12 15:14:05 by jjaniec           #+#    #+#             */
-/*   Updated: 2018/06/20 15:07:08 by jjaniec          ###   ########.fr       */
+/*   Updated: 2018/06/20 17:05:12 by jjaniec          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -100,7 +100,7 @@ t_lexeme				*create_lexeme(size_t type, char *data, \
 size_t					get_lexeme_type(char *s, int *pos, \
 							char **data, size_t *type_details);
 
-int						lexeme_type_word(char *s, int *pos, char **data, \
+size_t					lexeme_type_word(char *s, int *pos, char **data, \
 							int *env_assigns_passed);
 
 #endif
